+++ playback/src/audio_backend/rodio.rs
@@ -5,8 +5,9 @@ use cpal::traits::{DeviceTrait, HostTrait};
 use thiserror::Error;
 
 use super::Sink;
-use crate::audio::{convert, AudioPacket};
 use crate::config::AudioFormat;
+use crate::convert;
+use crate::decoder::AudioPacket;
 use crate::player::{NUM_CHANNELS, SAMPLE_RATE};
 
 #[cfg(all(